requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-jazzmin==2.6.0
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.6.0
